+++ smack-community-users.php
@@ -131,7 +131,7 @@ include 'core/sidebar.php';
     </div>
 
     <?php if (isset($msg)): ?>
-        <div class="msg">> <?php echo htmlspecialchars($msg); ?></div>
+        <div class="alert alert-success">&gt; <?php echo htmlspecialchars($msg); ?></div>
     <?php endif; ?>
 
     <div class="box">